fixes the blank search table when browsing lots of files

+++ private-components/swingui/src/main/java/org/limewire/ui/swing/search/SponsoredResultsPanel.java
@@ -53,7 +53,7 @@ public class SponsoredResultsPanel extends JXPanel {
         this.storeMediator = storeMediator;
         this.homeMediator = homeMediator;
         setLayout(new VerticalLayout(10));
-        setMaximumSize(new Dimension(areaWidth, Short.MAX_VALUE));
+        setMaximumSize(new Dimension(areaWidth, Integer.MAX_VALUE));
         setMinimumSize(new Dimension(areaWidth, Short.MAX_VALUE));
     }
